It works. In an awkward sorta way
