+++ autots/__init__.py
@@ -4,10 +4,11 @@ Automated Time Series Model Selection for Python
 https://github.com/winedarksea/AutoTS
 """
 from autots.datasets import load_toy_daily
+from autots.evaluator.auto_ts import AutoTS
 
 __version__ = '0.0.1'
 
-__all__ = ['load_toy_daily']
+__all__ = ['load_toy_daily', 'AutoTS']
 
 # import logging
 # logger = logging.getLogger(__name__)
